futurix-ai/extraction-service/requirements.txt
fastapi
uvicorn[standard]
pytesseract
python-multipart
Pillow
pydantic-settings
pytest
httpx
